integrate "inlineswap" (a non-invasive HTML preprocessor)
inlineswap is a super-simple HTML preprocessor that allows you to
automaticall update parts of HTML pages in-place.  For example, if you
change the header in index.html, it can automatically change the header
in all of the other pages for you.  See the "<-- header -->" command in
index.html and the "<-- get header from index.html -->" commands in
other pages.

This is purely for convenience, and if you don't like this, you can
simply ignore or delete the inlineswap commands (which are just HTML
comments anyway) and edit the files manually.

To apply the updates with inlineswap, run ./build.sh from the root
directory of the project.